- architecture cleanup - separates onboarding screen and deck screen for year in review - creates a common scaffold for year in review screen - refactors showing survey dialog logic
